# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SupportNova API",
    "description": "Complaint intelligence: GenAI analysis + Python ground-truth validation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "storefront"
        ],
        "summary": "List Products",
        "operationId": "list_products",
        "parameters": [
          {
            "name": "product_line",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Line"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search name, description, specs",
              "title": "Q"
            },
            "description": "Search name, description, specs"
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Price"
            }
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Price"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "featured",
                "price_asc",
                "price_desc",
                "newest",
                "best_selling"
              ],
              "type": "string",
              "default": "featured",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductOut"
                  },
                  "title": "Response List Products"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{sku}": {
      "get": {
        "tags": [
          "storefront"
        ],
        "summary": "Get Product",
        "operationId": "get_product",
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checkout": {
      "post": {
        "tags": [
          "storefront"
        ],
        "summary": "Place Order",
        "operationId": "place_order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShopOrderOut"
                  },
                  "type": "array",
                  "title": "Response Place Order"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/orders": {
      "get": {
        "tags": [
          "storefront"
        ],
        "summary": "My Shop Orders",
        "operationId": "my_shop_orders",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShopOrderOut"
                  },
                  "type": "array",
                  "title": "Response My Shop Orders"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{order_ref}/simulate": {
      "post": {
        "tags": [
          "storefront"
        ],
        "summary": "Simulate Delivery",
        "operationId": "simulate_delivery",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Ref"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/product-images/{image_id}": {
      "get": {
        "tags": [
          "storefront"
        ],
        "summary": "Product Image",
        "operationId": "product_image",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/*": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/products": {
      "get": {
        "tags": [
          "storefront"
        ],
        "summary": "Admin List Products",
        "operationId": "admin_list_products",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductOut"
                  },
                  "type": "array",
                  "title": "Response Admin List Products"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "storefront"
        ],
        "summary": "Create Product",
        "operationId": "create_product",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/products/{sku}": {
      "patch": {
        "tags": [
          "storefront"
        ],
        "summary": "Update Product",
        "operationId": "update_product",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/products/{sku}/images": {
      "post": {
        "tags": [
          "storefront"
        ],
        "summary": "Upload Product Image",
        "operationId": "upload_product_image",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_product_image"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/products/{sku}/images/order": {
      "put": {
        "tags": [
          "storefront"
        ],
        "summary": "Reorder Product Images",
        "operationId": "reorder_product_images",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageOrderIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/products/{sku}/images/{image_id}": {
      "delete": {
        "tags": [
          "storefront"
        ],
        "summary": "Delete Product Image",
        "operationId": "delete_product_image",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku"
            }
          },
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutIn": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/CheckoutLineIn"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Lines"
          },
          "shipping_method": {
            "type": "string",
            "enum": [
              "standard",
              "express"
            ],
            "title": "Shipping Method",
            "default": "standard"
          }
        },
        "type": "object",
        "required": [
          "lines"
        ],
        "title": "CheckoutIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProductCreate": {
        "properties": {
          "sku": {
            "type": "string",
            "pattern": "^[A-Z0-9][A-Z0-9-]{2,39}$",
            "title": "Sku",
            "description": "e.g. VH-PHN-NX6"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Name"
          },
          "product_line": {
            "type": "string",
            "title": "Product Line"
          },
          "price": {
            "type": "number",
            "maximum": 100000.0,
            "exclusiveMinimum": 0.0,
            "title": "Price"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Description"
          },
          "specs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Specs"
          }
        },
        "type": "object",
        "required": [
          "sku",
          "name",
          "product_line",
          "price",
          "description"
        ],
        "title": "ProductCreate"
      },
...
      "ProductOut": {
        "properties": {
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "product_line": {
            "type": "string",
            "title": "Product Line"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "specs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Specs"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ProductImageOut"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "sku",
          "name",
          "product_line",
          "price",
          "description",
          "specs",
          "is_active",
          "images"
        ],
        "title": "ProductOut"
      },
      "ProductUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "product_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Line"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "specs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 12
              },
              {
                "type": "null"
              }
            ],
            "title": "Specs"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "ProductUpdate"
      },
...
      "ShopOrderOut": {
        "properties": {
          "order_ref": {
            "type": "string",
            "title": "Order Ref"
          },
          "checkout_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout Ref"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "product_category": {
            "type": "string",
            "title": "Product Category"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "shipping_method": {
            "type": "string",
            "title": "Shipping Method"
          },
          "order_date": {
            "type": "string",
            "format": "date",
            "title": "Order Date"
          },
          "committed_delivery_date": {
            "type": "string",
            "format": "date",
            "title": "Committed Delivery Date"
          },
          "delivered_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "order_ref",
          "checkout_ref",
          "product_name",
          "product_category",
          "quantity",
          "amount",
          "shipping_method",
          "order_date",
          "committed_delivery_date",
          "delivered_date",
          "status"
        ],
        "title": "ShopOrderOut"
      },
      "SimulateIn": {
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "on_time",
              "late",
              "lost",
              "damaged"
            ],
            "title": "Outcome"
          },
          "days": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 0.0,
            "title": "Days",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "outcome"
        ],
        "title": "SimulateIn"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Clerk session token",
        "scheme": "bearer"
      }
    }
  }
}